Issue #12 refactor: toStringDateFormat 위치 변경
