+++ app/src/main/java/com/drunkenboys/calendarun/util.kt
@@ -1,4 +1,4 @@
-package com.drunkenboys.calendarun.data.calendar
+package com.drunkenboys.calendarun
 
 import android.annotation.SuppressLint
 import java.text.SimpleDateFormat
